Added Search Page connection
Fixed color complexity
